work

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,11 +7,11 @@
  */
 int main(void)
 {
- 	unsigned long long i = 0, j = 1, k, tmp;
+	unsigned long i = 0, j = 1, k, tmp;
 
 	for (k = 0; k < 50; k++)
 	{
-		printf("%llu, ", i);
+		printf("%lu, ", i);
 		tmp = j;
 		j = i + j;
 		i = tmp;
